feat: 优化协作管理路径
