+++ web/src/management/pages/list/components/SpaceModify.vue
@@ -48,9 +48,11 @@ import { useStore } from 'vuex'
 import { pick as _pick } from 'lodash-es'
 import { ElMessage } from 'element-plus'
 import 'element-plus/theme-chalk/src/message.scss'
+
 import { QOP_MAP } from '@/management/utils/constant'
-import MemberSelect from './MemberSelect.vue'
 import { type IMember, type IWorkspace, UserRole } from '@/management/utils/types/workSpace'
+
+import MemberSelect from '@/management/components/CooperModify/MemberSelect.vue'
 
 const store = useStore()
 const emit = defineEmits(['on-close-codify', 'onFocus', 'change', 'blur'])
