+++ web/src/management/styles/edit-btn.scss
@@ -15,4 +15,10 @@
   .btn-txt {
     font-size: 12px;
   }
+
+  .view-icon {
+    font-size: 20px;
+    height: 29px;
+    line-height: 29px;
+  }
 }
